# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Little Bear RAG Backend API
  version: 0.1.0
  description: >
    P0 OpenAPI contract for Little Bear RAG backend. The contract covers setup,
    authentication, organization, permission, knowledge base, document import,
    indexing operations, query, audit, healthcheck, config, and internal model
    gateway APIs. P0 uses external embedding, rerank, and LLM provider
    endpoints instead of built-in model simulators, and does not include complete
    config approval or rollback workflows.

    Backend contract paths use /internal/v1. OpenAPI, backend routes, frontend
    SDKs, and contract tests should all use this path prefix consistently.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /internal/v1/sessions:
    post:
      tags: [Auth]
      summary: Create session
      operationId: createSession
      security: []
      x-token-type: anonymous
      x-audit-required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/sessions/current:
    delete:
      tags: [Auth]
      summary: Revoke current session
      operationId: deleteCurrentSession
      x-token-type: access
      x-required-scope: auth:session
      x-audit-required: true
      responses:
        "204":
          description: Session revoked
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/token-refreshes:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      operationId: createTokenRefresh
      security:
        - bearerAuth: []
      x-token-type: refresh
      x-audit-required: true
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/users/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getCurrentUser
      x-token-type: access
      x-required-scope: auth:session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/users/me/password:
    put:
      tags: [Auth]
      summary: Change current user password
      operationId: putCurrentUserPassword
      x-token-type: access
      x-required-scope: auth:password:update:self
      x-audit-required: true
      x-confirmation-required: true
      x-confirmation-condition: always
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        "204":
          description: Password changed
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RequestId:
      type: string
      examples: ["req_001"]
...
    AnyObject:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [request_id, error_code, message, retryable]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        error_code:
          type: string
        message:
          type: string
        stage:
          type: string
        retryable:
          type: boolean
        details:
          $ref: "#/components/schemas/AnyObject"
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        enterprise_code:
          type: string
          description: 多企业存在同名账号时必填；单企业部署可省略。
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
    PasswordChangeRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, disabled, locked, deleted]
        departments:
          type: array
          items:
            $ref: "#/components/schemas/Department"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        scopes:
          type: array
          items:
            type: string
...
    UserResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          $ref: "#/components/schemas/User"
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [active, disabled, deleted]
        is_primary:
          type: boolean
          default: false
        is_default:
          type: boolean
          default: false
...
    Role:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        scope_type:
          type: string
          enum: [enterprise, department, knowledge_base]
        is_builtin:
          type: boolean
        status:
          type: string
          enum: [active, disabled, archived]
        scopes:
          type: array
          items:
            type: string
...